Move memo to its own module under note_encryption, turn try_from into a match cmp
